Proceed with page setup even if login check fails
If the API request to check the login status fails (for instance,
because the response contains PHP warnings that cause the JSON parsing
to fail), we can still proceed loading the page. This is mainly useful
for local development, so that it’s possible to work on the frontend
without setting up OAuth (and the whole /data/project directory
structure).