Add qualifier support to CSV format
I don’t fully understand how this works, given that at import time we
don’t yet have a statement ID to refer to, but just doing the same thing
as the v1 import seems to work.

I considered requiring “qalP1234” instead of “qal1234”, which would
slightly simplify the code and also reduce reliance on partial entity
IDs (the full entity ID of the property is “P1234”, with the P), but
decided against it, both for aesthetic and compatibility reasons.