Fix bug in adding references in CSV format
$lastSources is a reference (in the PHP sense – &value) so that adding
further snaks to the previous source (with s123) works, which means that
any time we assign to it, we must first break the reference using
unset() (compare L1058 and L1103). This was missing in the beginning of
the loop body, with the consequence that in a multi-line import, the
last reference of each line except the last would be broken (it would
have been reset to null by the next line).