Don’t try to send incomplete values to the API
If parseValueV1() can’t parse the value, the $cmd it returns will still
have a 'datavalue', but there won’t be a 'value' inside it (but rather a
'text', along with 'type'=>'unknown'). This is apparently not checked at
any intervening point (neither in the V1 nor CSV formats), and
eventually the commandAdd*() methods would try to send the incomplete
data to the API, after producing several PHP warnings due to the missing
value (several of them inside getSnakType()), which would also break the
JSON response as usual.

Work around this by checking if we have a value just before sending the
API request. This means that the bad values will still be shown in the
batch view in a weird format (raw JSON), but I don’t know how that
should be fixed instead. This at least fixes the server-side warnings
and makes the batch not stuck indefinitely.